Setting av hendelsestatus

+++ sendsoknad-business/src/main/java/no/nav/sbl/dialogarena/soknadinnsending/business/db/soknad/SoknadRepositoryJdbc.java
@@ -208,6 +208,14 @@ public class SoknadRepositoryJdbc extends NamedParameterJdbcDaoSupport implement
         return soknad;
     }
 
+    public WebSoknad hentSoknadMedVedlegg(Long id) {
+        WebSoknad soknad = hentSoknad(id);
+        if (soknad != null && soknad.getBrukerBehandlingId() != null) {
+            leggTilBrukerdataOgVedleggPaaSoknad(soknad, soknad.getBrukerBehandlingId());
+        }
+        return soknad;
+    }
+
 
     private void leggTilBrukerdataOgVedleggPaaSoknad(WebSoknad soknad, String behandlingsId) {
         soknad.medBrukerData(hentAlleBrukerData(behandlingsId)).medVedlegg(vedleggRepository.hentVedlegg(behandlingsId));
@@ -504,45 +512,63 @@ public class SoknadRepositoryJdbc extends NamedParameterJdbcDaoSupport implement
         ids.forEach(id -> slettSoknadPermanent(id, HendelseType.PERMANENT_SLETTET_AV_SYSTEM));
     }
 
-    public void slettGamleIkkeInnsendteSoknader(int dager) {
+    public List<WebSoknad> slettGamleIkkeInnsendteSoknader(int dager) {
         String sql = "select soknad_id from soknad where status=? and opprettetdato < sysdate - interval '"+ dager + "' day";
 
         List<Long> ids = getJdbcTemplate().queryForList(sql, Long.class, SoknadInnsendingStatus.UNDER_ARBEID.name());
 
         logger.info("Fant {} soknader eldre enn {} dager. Sletter soknader med disse ids: {}", ids.size(), dager, ids);
-        ids.forEach(id -> slettSoknad(id, SoknadInnsendingStatus.AVBRUTT_AUTOMATISK));
+        List<WebSoknad> slettedeSoknader = new LinkedList<>();
+        ids.forEach(id -> slettedeSoknader.add(slettSoknad(id, SoknadInnsendingStatus.AVBRUTT_AUTOMATISK)));
+        return slettedeSoknader;
     }
 
-    public void slettGamleSoknaderPermanent(int dager) {
+    public List<WebSoknad> slettGamleSoknaderPermanent(int dager) {
         String sql = "select soknad_id from soknad where opprettetdato <  sysdate - interval '"+ dager + "' day(3)";
 
         List<Long> ids = getJdbcTemplate().queryForList(sql, Long.class);
 
         logger.info("Fant {} soknader eldre enn {} dager. Sletter soknader permanent med disse ids: {}", ids.size(), dager, ids);
-        ids.forEach(id -> slettSoknadPermanent(id, HendelseType.PERMANENT_SLETTET_AV_SYSTEM));
+        List<WebSoknad> slettedeSoknader = new LinkedList<>();
+        ids.forEach(id -> slettedeSoknader.add(slettSoknadPermanent(id, HendelseType.PERMANENT_SLETTET_AV_SYSTEM)));
+        return slettedeSoknader;
     }
 
-    private void slettSoknad(long soknadId, SoknadInnsendingStatus status) {
-        logger.debug("Sletter søknad med ID: " + soknadId);
+    private WebSoknad slettSoknad(long soknadId, SoknadInnsendingStatus status) {
+        logger.info("SoknadId {}: Skal slette søknad",  soknadId);
+        WebSoknad soknad = hentSoknadMedVedlegg(soknadId);
+        if (soknad == null || soknad.getBrukerBehandlingId() == null) {
+            logger.warn("soknadId {}: ikke funnet søknad eller behandlingsid er null");
+            throw new RuntimeException("soknadId {}: ikke funnet søknad eller behandlingsid er null");
+        }
+        logger.info("{}: Hentet soknadId = {} soknad for sletting", soknad.getBrukerBehandlingId(), soknad.getSoknadId());
         getJdbcTemplate().update("delete from faktumegenskap where soknad_id = ?", soknadId);
         getJdbcTemplate().update("delete from soknadbrukerdata where soknad_id = ?", soknadId);
         getJdbcTemplate().update("update vedlegg set data = null, storrelse = 0 where soknad_id = ?", soknadId);
         getJdbcTemplate().update("update soknad set status=?, sistlagret = CURRENT_TIMESTAMP where soknad_id = ?", status.name(), soknadId);
+        hendelseRepository.registrerHendelse(soknad, HendelseType.AVBRUTT_AUTOMATISK);
+        logger.info("{}: Søknad slettet",  soknad.getBrukerBehandlingId());
+        return soknad;
     }
 
-    public void slettSoknadPermanent(long soknadId, HendelseType aarsakTilSletting) {
-        logger.debug("Permanent sletting av søknad med ID: " + soknadId);
-        WebSoknad soknad = hentSoknad(soknadId);
+    public WebSoknad slettSoknadPermanent(long soknadId, HendelseType aarsakTilSletting) {
+        logger.debug("SoknadId {}: Skal permanent slette søknad", soknadId);
+        WebSoknad soknad = hentSoknadMedVedlegg(soknadId);
+        if (soknad == null || soknad.getBrukerBehandlingId() == null) {
+            logger.warn("soknadId {}: ikke funnet søknad eller behandlingsid er null");
+            throw new RuntimeException("soknadId {}: ikke funnet søknad eller behandlingsid er null");
+        }
         getJdbcTemplate().update("delete from faktumegenskap where soknad_id = ?", soknadId);
         getJdbcTemplate().update("delete from soknadbrukerdata where soknad_id = ?", soknadId);
         getJdbcTemplate().update("delete from vedlegg where soknad_id = ?", soknadId);
         getJdbcTemplate().update("delete from soknad where soknad_id = ?", soknadId);
         hendelseRepository.registrerHendelse(soknad, aarsakTilSletting);
+        return soknad;
     }
 
     public void slettSoknad(WebSoknad soknad, HendelseType aarsakTilSletting) {
         if (aarsakTilSletting == HendelseType.AVBRUTT_AV_BRUKER) {
-            slettSoknadPermanent(soknad.getSoknadId(),  HendelseType.AVBRUTT_AV_BRUKER);
+            slettSoknadPermanent(soknad.getSoknadId(),  HendelseType.PERMANENT_SLETTET_AV_BRUKER);
         } else {
             slettSoknad(soknad.getSoknadId(), convertStatus(aarsakTilSletting));
             hendelseRepository.registrerHendelse(soknad, aarsakTilSletting);
